Added custom work sharing implementatation
Kind of ugly but performs better than messaging implementaitons.
we attempted in the past.  For each iteration we use a single
shared vector containing the random arrangement of nodeids
each of the nthreads we use then runs a loop while there are nodes
in this vector.  For each iteration of the loop it'll pop
off the last node and then run an update step using this node
for it's walk.

+++ src/config.rs
@@ -13,7 +13,7 @@ pub struct Config {
     input_file: String,
     perf_file: String,
     weight_file: String,
-    nthreads: u32,
+    nthreads: usize,
 }
 
 impl Config {
@@ -48,7 +48,7 @@ impl Config {
     pub fn weight_file(&self) -> &str {
         &self.weight_file[..]
     }
-    pub fn nthreads(&self) -> u32 {
+    pub fn nthreads(&self) -> usize {
         self.nthreads
     }
 }
